+++ main_module/src/api/Api_resQuestions.cpp
@@ -1,22 +1,39 @@
 #include "Api_resQuestions.h"
 
-Api_resQuestions::Api_resQuestions(ResourceQuestions& resQuestions) : resQuestions(resQuestions) {};
+Api_resQuestions::Api_resQuestions(ResourceQuestions& resQuestions, PermissionChecker& permChecker) : resQuestions(resQuestions), permChecker(permChecker) {};
 
-void Api_resQuestions::get_all(const httplib::Request&, httplib::Response& res) {
+void Api_resQuestions::get_all(const httplib::Request& req, httplib::Response& res) {
+	PermissionInfo p_info = permChecker.check(req, "quest:list:read");
+	if (p_info.status==401 || p_info.status==418) {
+		res.status = p_info.status; return;
+	}
+
 	auto questions = resQuestions.get_all();
 	nlohmann::json json_res = nlohmann::json::array();
 	for (const auto& q : questions) {
-		json_res.push_back(
-			{ {"name", q.name},{"version", q.version},{"author_id", q.author_id} }
-		);
+		if (p_info.status==200 || p_info.user_id == q.author_id){
+			json_res.push_back(
+				{ {"name", q.name},{"version", q.version},{"author_id", q.author_id} }
+			);
+		}
 	}
 	res.set_content(json_res.dump(), "application/json");
 }
 
 void Api_resQuestions::get_info(const httplib::Request& req, httplib::Response& res) {
+	PermissionInfo p_info = permChecker.check(req, "quest:read");
+	if (p_info.status==401 || p_info.status==418) {
+		res.status = p_info.status; return;
+	}
 	int question_id = std::stoi(req.matches[1]);
 	int version = std::stoi(req.get_param_value("version"));
 	QuestionInfo info = resQuestions.get_info(question_id, version);
+	if (p_info.status==403 && 
+		(info.author_id != p_info.user_id && (!resQuestions.check_presence(question_id, p_info.user_id)))){
+		res.status = p_info.status; return;
+	}
+
+
 	nlohmann::json json_res = {
 		{"name", info.name},
 		{"text", info.text},
@@ -27,27 +44,48 @@ void Api_resQuestions::get_info(const httplib::Request& req, httplib::Response& 
 }
 
 void Api_resQuestions::update_question(const httplib::Request& req, httplib::Response& res) {
+	PermissionInfo p_info = permChecker.check(req, "quest:update");
+	if (p_info.status==401 || p_info.status==418) {
+		res.status = p_info.status; return;
+	}
 	int question_id = std::stoi(req.matches[1]);
+	QuestionInfo info = resQuestions.get_info(question_id, 1);
+	int author_id = info.author_id;
+	if (p_info.status==403 && author_id != p_info.user_id){
+		res.status = p_info.status; return;
+	}
+
 	std::string name = req.get_param_value("name");
 	std::string text = req.get_param_value("text");
 	nlohmann::json options = nlohmann::json::parse(req.get_param_value("options"));
 	int correct_option = std::stoi(req.get_param_value("correct_option"));
-	int author_id = std::stoi(req.get_param_value("author_id"));
 	resQuestions.update_question(question_id, name, text, options, correct_option, author_id);
 }
 
 
 void Api_resQuestions::create_question(const httplib::Request& req, httplib::Response& res) {
+	PermissionInfo p_info = permChecker.check(req, "quest:create");
+	if (p_info.status != 200) {
+		res.status = p_info.status; return;
+	}
+	int author_id = p_info.user_id;
 	std::string name = req.get_param_value("name");
 	std::string text = req.get_param_value("text");
 	nlohmann::json options = nlohmann::json::parse(req.get_param_value("options"));
 	int correct_option = std::stoi(req.get_param_value("correct_option"));
-	int author_id = std::stoi(req.get_param_value("author_id"));
 	int new_question_id = resQuestions.create_question(name, text, options, correct_option, author_id);
 	res.set_content(std::to_string(new_question_id), "text/plain");
 }
 
 void Api_resQuestions::delete_question(const httplib::Request& req, httplib::Response& res) {
+	PermissionInfo p_info = permChecker.check(req, "quest:del");
+	if (p_info.status==401 || p_info.status==418) {
+		res.status = p_info.status; return;
+	}
 	int question_id = std::stoi(req.matches[1]);
+	QuestionInfo info = resQuestions.get_info(question_id, 1);
+	if (p_info.status==403 && info.author_id != p_info.user_id){
+		res.status = p_info.status; return;
+	}
 	resQuestions.delete_question(question_id);
 }
